add fuction deletePerson
