+++ sprint1/pwz/learn_java_basic/more_about_java_oop/TryJavaOOP_Interface.java
